Added support to customise bold headings of org levels and priority items
Spacemacs theme makes headings of org items level 1 and 2 inherit bold weight without apparent way to customise it.
The same goes for agenda (priority) items, which are bold.

This commit adds two custom variables:

- spacemacs-theme-org-bold
controls if level 1 and 2 items should be bold or not
default: t

- spacemacs-theme-org-priority-bold
controls if items in the agenda should be bold
default: t